+++ app/code/Magento/Usa/Model/Shipping/Carrier/Fedex.php
@@ -174,7 +174,7 @@ class Fedex
      *
      * @param string $wsdl
      * @param bool|int $trace
-     * @return SoapClient
+     * @return \SoapClient
      */
     protected function _createSoapClient($wsdl, $trace = false)
     {
@@ -190,7 +190,7 @@ class Fedex
     /**
      * Create rate soap client
      *
-     * @return SoapClient
+     * @return \SoapClient
      */
     protected function _createRateSoapClient()
     {
@@ -200,7 +200,7 @@ class Fedex
     /**
      * Create ship soap client
      *
-     * @return SoapClient
+     * @return \SoapClient
      */
     protected function _createShipSoapClient()
     {
@@ -210,7 +210,7 @@ class Fedex
     /**
      * Create track soap client
      *
-     * @return SoapClient
+     * @return \SoapClient
      */
     protected function _createTrackSoapClient()
     {