+++ dev/tests/static/testsuite/Magento/Test/Integrity/ReferenceBlacklist.php
@@ -1,0 +1,49 @@
+<?php
+/**
+ * Legacy class usages excluded from the ClassesTest->testClassReferences for PSR-X standards
+ *
+ * {license_notice}
+ *
+ * @category    tests
+ * @package     static
+ * @subpackage  Integrity
+ * @copyright   {copyright}
+ * @license     {license_link}
+ */
+
+return array(
+    'PEAR_PackageFile',
+    'Simple_Xml',
+    'PersistentCustomerSegmentation',
+    'SolrClient',
+    'SolrQuery',
+    'RegionUpdater', // JavaScript usage
+    'Fixture_Module', // JavaScript usage
+    'TheCompoundNamespace_TheCompoundModule',
+    'Some_Module',
+    'Not_Existed_Class',
+    'CustomSelect',
+
+    // found in /dev/tests/unit/testsuite/Magento/Test/Tools/Di/_files/app/code/Magento/SomeModule/Model/Test.php
+    '\Magento\SomeModule\Model\Element\Proxy',
+
+    'Map_Module',
+    'Module_One',
+    'Module_Two',
+    'Module_Name',
+    'Test',
+    'Local_Module',
+
+    // found in /dev/tests/unit/testsuite/Magento/Webapi/Controller/Soap/HandlerTest.php
+    '\Magento\Webapi\Controller\Soap\Security',
+
+    'Mage',
+    'Pear_Package_Parser_v2',
+    'PEAR_Config',
+    'PEAR_DependencyDB',
+    'PEAR_Frontend',
+    'PEAR_Command',
+    'Zend',
+    'PEAR_PackageFileManager2',
+    'PEAR',
+);