+++ CHANGELOG.markdown
@@ -33,6 +33,7 @@
   * Fixed Fatal error when SID presents in URL
   * Fixed absence of selection of a role assigned to an admin user
   * Fixed Oauth consumer credentials expiry not being correctly calculated and added credentials HTTP post to the consumer endpoint
+  * Fixed Namespace class references
 
 2.0.0.0-dev46
 =============